+++ src/Benchmarks/ArgumentsBenchmarks.cs
@@ -55,7 +55,7 @@ namespace Benchmarks
 
             for (int i = 0; i < 10000; i++)
             {
-                ExpressionUtil.ChangeType(variables["names"], typeof(InputType), null, null);
+                ExpressionUtil.ConvertObjectType(variables["names"], typeof(InputType), null, null);
             }
         }
 
@@ -75,7 +75,7 @@ namespace Benchmarks
 
             for (int i = 0; i < 10000; i++)
             {
-                ExpressionUtil.ChangeType(variables["names"], typeof(List<InputType>), null, null);
+                ExpressionUtil.ConvertObjectType(variables["names"], typeof(List<InputType>), null, null);
             }
         }
     }